cleanup in home page
