+++ dashboard-ui/src/components/AggregateResults/statistics.jsx
@@ -2,12 +2,31 @@ import { isDefined } from "./DataFunctions";
 
 function getBoxWhiskerData(lst) {
     const cleanLst = getCleanLst(lst);
-    // returns [min, q1, q3, max, median] for a list of data
     cleanLst.sort((a, b) => a - b);
-    const med = getMedian(cleanLst);
+
+    // Calculate median, Q1, Q3
     const q1 = getMedian(cleanLst.slice(0, Math.floor(cleanLst.length / 2)));
     const q3 = getMedian(cleanLst.slice(Math.floor(cleanLst.length / 2)));
-    return [Math.min(...cleanLst), q1, q3, Math.max(...cleanLst), med];
+
+    // Calculate Interquartile Range
+    const iqr = q3 - q1;
+
+    // Define outlier bounds
+    const lowerBound = q1 - 1.5 * iqr;
+    const upperBound = q3 + 1.5 * iqr;
+
+    // Filter out outliers
+    const filteredData = cleanLst.filter(value => value >= lowerBound && value <= upperBound);
+
+    // Recalculate the box plot values with the filtered data
+    const newMin = Math.min(...filteredData);
+    const newMax = Math.max(...filteredData);
+    const newQ1 = getMedian(filteredData.slice(0, Math.floor(filteredData.length / 2)));
+    const newQ3 = getMedian(filteredData.slice(Math.floor(filteredData.length / 2)));
+    const newMed = getMedian(filteredData);
+
+    // Return the updated box whisker data
+    return [newMin, newQ1, newQ3, newMax, newMed];
 }
 
 function getMedian(lst) {
@@ -129,6 +148,34 @@ const getMeanAcrossAll = (obj, keys = 'all') => {
     return getMean(data);
 };
 
+const getMeanAcrossAllWithoutOutliers = (obj, keys = 'all') => {
+    const data = [];
+    if (obj != undefined) {
+        for (const key of Object.keys(obj)) {
+            if (keys === 'all' || keys.includes(key)) {
+                data.push(...obj[key]);
+            }
+        }
+    }
+    const cleanLst = getCleanLst(data);
+    cleanLst.sort((a, b) => a - b);
+
+    // Calculate median, Q1, Q3
+    const q1 = getMedian(cleanLst.slice(0, Math.floor(cleanLst.length / 2)));
+    const q3 = getMedian(cleanLst.slice(Math.floor(cleanLst.length / 2)));
+
+    // Calculate Interquartile Range
+    const iqr = q3 - q1;
+
+    // Define outlier bounds
+    const lowerBound = q1 - 1.5 * iqr;
+    const upperBound = q3 + 1.5 * iqr;
+
+    // Filter out outliers
+    const filteredData = cleanLst.filter(value => value >= lowerBound && value <= upperBound);
+    return filteredData.reduce((a, b) => a + b, 0) / filteredData.length;
+};
+
 const getSeAcrossAll = (obj, keys = 'all') => {
     const data = [];
     if (obj != undefined) {
@@ -153,10 +200,11 @@ const calculateBestFitLine = (data) => {
     const b = (sumY - m * sumX) / n;
 
     const lineData = [
+        { x: 0, y: b },
         { x: Math.min(...data.map((x) => x.x)), y: m * Math.min(...data.map((x) => x.x)) + b },
         { x: Math.max(...data.map((x) => x.x)), y: m * Math.max(...data.map((x) => x.x)) + b },
     ];
     return lineData;
 };
 
-export { getBoxWhiskerData, getMean, getMedian, getMode, getStandDev, getStandardError, getMeanAcrossAll, getSeAcrossAll, calculateBestFitLine, getLogisticData };
+export { getBoxWhiskerData, getMean, getMedian, getMode, getStandDev, getStandardError, getMeanAcrossAll, getSeAcrossAll, calculateBestFitLine, getLogisticData, getMeanAcrossAllWithoutOutliers };
