+++ dashboard-ui/src/css/aggregateResults.css
@@ -47,6 +47,17 @@
     margin-bottom: 8px;
 }
 
+.side-by-side-plots {
+    display: flex;
+    flex-direction: row;
+    padding: 12px;
+    width: 1423px !important;
+}
+
+.side-by-side-plots>* {
+    width: 50%;
+}
+
 .outlinedPlot h4 {
     text-align: center;
     margin-bottom: 12px;
@@ -154,6 +165,10 @@
         min-width: 712px;
     }
 
+    .side-by-side-plots {
+        flex-direction: column;
+        align-items: center;
+    }
 }
 
 .delegationChoices>*, .side-by-side>* {
